// src/types/RecievedMessages.ts
export type RecievedMessage = {
    _data: {
      id: {
        fromMe: boolean;
        remote: string;
        id: string;
        _serialized: string;
      };
      viewed: boolean;
      body: string;
      type: string;
      t: number;
      notifyName: string;
      from: string;
      to: string;
      author: string;
      ack: number;
      invis: boolean;
      isNewMsg: boolean;
      star: boolean;
      kicNotified: boolean;
      recvFresh: boolean;
      isFromTemplate: boolean;
      thumbnail: string;
      pollInvalidated: boolean;
      isSentCagPollCreation: boolean;
      latestEditMsgKey: null | string;
      latestEditSenderTimestampMs: null | number;
      quotedMsg: {
        id: object;
        rowId: number;
        type: string;
        t: number;
        from: string;
        to: string;
        ack: number;
        invis: boolean;
        directPath: string;
        mimetype: string;
        duration: string;
        filehash: string;
        size: number;
        streamingSidecar: object;
        mediaKey: string;
        mediaKeyTimestamp: number;
        isViewOnce: boolean;
        width: number;
        height: number;
        interactiveAnnotations: any[];
        messageRangeIndex: string;
        deprecatedMms3Url: string;
        encFilehash: string;
        hasReaction: boolean;
        botTargetSenderJid: null | string;
        bizBotType: null | string;
        eventInvalidated: boolean;
        msgRowOpaqueData: object;
        staticUrl: string;
        pollInvalidated: boolean;
        star: boolean;
        isMdHistoryMsg: boolean;
        body: string;
        pollOptions: any[];
      };
      quotedStanzaID: string;
      quotedParticipant: string;
      mentionedJidList: any[];
      groupMentions: any[];
      isEventCanceled: boolean;
      eventInvalidated: boolean;
      isVcardOverMmsDocument: boolean;
      hasReaction: boolean;
      ephemeralDuration: number;
      ephemeralSettingTimestamp: number;
      disappearingModeInitiator: string;
      disappearingModeTrigger: string;
      productHeaderImageRejected: boolean;
      lastPlaybackProgress: number;
      isDynamicReplyButtonsMsg: boolean;
      isCarouselCard: boolean;
      parentMsgId: null | string;
      isMdHistoryMsg: boolean;
      stickerSentTs: number;
      isAvatar: boolean;
      lastUpdateFromServerTs: number;
      invokedBotWid: null | string;
      bizBotType: null | string;
      botResponseTargetId: null | string;
      botPluginType: null | string;
      botPluginReferenceIndex: null | string;
      botPluginSearchProvider: null | string;
      botPluginSearchUrl: null | string;
      botPluginMaybeParent: boolean;
      botReelPluginThumbnailCdnUrl: null | string;
      botMsgBodyType: null | string;
      requiresDirectConnection: null | boolean;
      bizContentPlaceholderType: null | string;
      hostedBizEncStateMismatch: boolean;
      senderOrRecipientAccountTypeHosted: boolean;
      placeholderCreatedWhenAccountIsHosted: boolean;
      links: any[];
    };
    mediaKey: undefined | string;
    id: {
      fromMe: boolean;
      remote: string;
      id: string;
      _serialized: string;
    };
    ack: number;
    hasMedia: boolean;
    body: string;
    type: string;
    timestamp: number;
    from: string;
    to: string;
    author: string;
    deviceType: string;
    isForwarded: undefined | boolean;
    forwardingScore: number;
    isStatus: boolean;
    isStarred: boolean;
    broadcast: undefined | boolean;
    fromMe: boolean;
    hasQuotedMsg: boolean;
    hasReaction: boolean;
    duration: undefined | number;
    location: undefined | string;
    vCards: any[];
    inviteV4: undefined | string;
    mentionedIds: any[];
    orderId: undefined | string;
    token: undefined | string;
    isGif: boolean;
    isEphemeral: undefined | boolean;
    links: any[];
  };